Time: 61 ms (85.85%) | Memory: 83.1 MB (47.74%) - LeetSync

+++ 2255-minimum-swaps-to-group-all-1s-together-ii/minimum-swaps-to-group-all-1s-together-ii.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     int minSwaps(vector<int>& nums) {
-        int k = accumulate(nums.begin(), nums.end(), 0);
-        int n = nums.size();
-        int cnt = accumulate(nums.begin(), nums.begin() + k, 0);
-        int mx = cnt;
+        long long k = accumulate(nums.begin(), nums.end(), 0);
+        long long n = nums.size();
+        long long cnt = accumulate(nums.begin(), nums.begin() + k, 0);
+        long long mx = cnt;
         for (int i = k; i < n + k; ++i) {
             cnt += nums[i % n] - nums[(i - k + n) % n];
             mx = max(mx, cnt);
